# Inline small scalars (up to 8 bytes) into ScalarAttribute instance