--- a/src/Fabulous.Benchmarks/Benchmarks.fs
+++ b/src/Fabulous.Benchmarks/Benchmarks.fs
@@ -52,10 +52,10 @@ module NestedTreeCreation =
 
 
 module DiffingAttributes =
-//    [<Struct>]
+    //    [<Struct>]
     type Model = { depth: int; counter: int }
 
-//    [<Struct>]
+    //    [<Struct>]
     type Msg = IncBy of int
 
     let update msg model =
@@ -120,11 +120,12 @@ module DiffingSmallScalars =
     let rec viewBoxedInner depth counter =
         // this is to emulate changing value only once per 5 updates
         let value = counter / 2UL
+
         Stack() {
             BoxedNumericBag(value, value, float value)
             BoxedNumericBag(value, value, float value)
             BoxedNumericBag(value, value, float value)
-            
+
             if (depth > 0) then
                 viewBoxedInner(depth - 1) counter
 
@@ -135,20 +136,24 @@ module DiffingSmallScalars =
     let rec viewInlineInner depth counter =
         // this is to emulate changing value only once per 5 updates
         let value = counter / 2UL
+
         Stack() {
             InlineNumericBag(value, value, float value)
             InlineNumericBag(value, value, float value)
             InlineNumericBag(value, value, float value)
-            
+
             if (depth > 0) then
                 viewInlineInner(depth - 1) counter
 
             if (depth > 0) then
                 viewInlineInner(depth - 2) counter
         }
-        
-    let viewBoxed model = viewBoxedInner model.depth model.counter
-    let viewInline model = viewInlineInner model.depth model.counter
+
+    let viewBoxed model =
+        viewBoxedInner model.depth model.counter
+
+    let viewInline model =
+        viewInlineInner model.depth model.counter
 
     [<MemoryDiagnoser>]
     [<SimpleJob(RuntimeMoniker.Net60)>]
@@ -158,13 +163,17 @@ module DiffingSmallScalars =
 
         [<Params(true, false)>]
         member val boxed = true with get, set
-        
+
         [<Benchmark>]
         member x.ProcessIncrements() =
             let program =
-                
-                let view = if x.boxed then viewBoxed else viewInline
-                
+
+                let view =
+                    if x.boxed then
+                        viewBoxed
+                    else
+                        viewInline
+
                 StatefulWidget.mkSimpleView(fun () -> { depth = x.depth; counter = 0UL }) update view
 
             let instance = Run.Instance program
@@ -178,14 +187,14 @@ module DiffingSmallScalars =
 
 [<EntryPoint>]
 let main argv =
-//    BenchmarkRunner.Run<NestedTreeCreation.Benchmarks>()
+    //    BenchmarkRunner.Run<NestedTreeCreation.Benchmarks>()
 //    |> ignore
 //
 //    BenchmarkRunner.Run<DiffingAttributes.Benchmarks>()
 //    |> ignore
-    
+
     printfn "Hello"
-    
+
     BenchmarkRunner.Run<DiffingSmallScalars.Benchmarks>()
     |> ignore