--- a/src/Fabulous.Tests/TestUI.Attributes.fs
+++ b/src/Fabulous.Tests/TestUI.Attributes.fs
@@ -2,6 +2,7 @@ module Tests.TestUI_Attributes
 
 open System
 open Fabulous
+open Fabulous.ScalarAttributeDefinitions
 open Tests.Platform
 
 module Attributes =