--- a/src/Fabulous.Tests/TestUI.ViewUpdaters.fs
+++ b/src/Fabulous.Tests/TestUI.ViewUpdaters.fs
@@ -18,18 +18,16 @@ let updateTextColor _ (newValueOpt: string voption) (node: IViewNode) =
 let updateAutomationId _ (newValueOpt: string voption) (node: IViewNode) =
     let el = node.Target :?> TestViewElement
     el.AutomationId <- ValueOption.defaultValue "" newValueOpt
-    
-    
+
+
 let updateNumericValueOne _ (newValueOpt: uint64 voption) (node: IViewNode) =
     let el = node.Target :?> TestNumericBag
     el.valueOne <- ValueOption.defaultValue 0UL newValueOpt
-    
+
 let updateNumericValueTwo _ (newValueOpt: uint64 voption) (node: IViewNode) =
     let el = node.Target :?> TestNumericBag
     el.valueTwo <- ValueOption.defaultValue 0UL newValueOpt
-    
+
 let updateNumericValueThree _ (newValueOpt: float voption) (node: IViewNode) =
     let el = node.Target :?> TestNumericBag
     el.valueThree <- ValueOption.defaultValue 0. newValueOpt
-    
-