… src/Fabulous.XamarinForms/Attributes.fs
… src/Fabulous.XamarinForms/Program.fs
… src/Fabulous.XamarinForms/Views/Cells/SwitchCell.fs
… src/Fabulous.XamarinForms/Views/Cells/ViewCell.fs
… src/Fabulous.XamarinForms/Views/Cells/_Cell.fs
… src/Fabulous.XamarinForms/Views/Collections/CarouselView.fs
… src/Fabulous.XamarinForms/Views/Collections/CollectionView.fs
… src/Fabulous.XamarinForms/Views/Collections/ListView.fs
… src/Fabulous.XamarinForms/Views/Collections/_ItemsView.fs
… src/Fabulous.XamarinForms/Views/Controls/Button.fs
… src/Fabulous.XamarinForms/Views/Controls/Entry.fs
… src/Fabulous.XamarinForms/Views/Controls/FormattedLabel.fs
… src/Fabulous.XamarinForms/Views/Controls/Image.fs
… src/Fabulous.XamarinForms/Views/Controls/IndicatorView.fs
… src/Fabulous.XamarinForms/Views/Controls/Label.fs
… src/Fabulous.XamarinForms/Views/Controls/RadioButton.fs
… src/Fabulous.XamarinForms/Views/Controls/Slider.fs
… src/Fabulous.XamarinForms/Views/Controls/Span.fs
… src/Fabulous.XamarinForms/Views/Controls/WebView.fs
… src/Fabulous.XamarinForms/Views/Layouts/Frame.fs
… src/Fabulous.XamarinForms/Views/Layouts/RelativeLayout.fs
… src/Fabulous.XamarinForms/Views/Layouts/ScrollView.fs
… src/Fabulous.XamarinForms/Views/Layouts/SwipeItem.fs
… src/Fabulous.XamarinForms/Views/Layouts/SwipeItems.fs
… src/Fabulous.XamarinForms/Views/Layouts/SwipeView.fs
… src/Fabulous.XamarinForms/Views/Layouts/_TemplatedView.fs
… src/Fabulous.XamarinForms/Views/Pages/FlyoutPage.fs
… src/Fabulous.XamarinForms/Views/Pages/NavigationPage.fs
… src/Fabulous.XamarinForms/Views/Shapes/Ellipse.fs
… src/Fabulous.XamarinForms/Views/Shapes/Geometries/GeometryGroup.fs