Detailed description and benchmark results are coming.

here is a preview

|            Method | depth | boxed |    Mean |    Error |   StdDev |       Gen 0 |      Gen 1 |      Gen 2 | Allocated |
|------------------ |------ |------ |--------:|---------:|---------:|------------:|-----------:|-----------:|----------:|
| ProcessIncrements |    15 | False | 1.162 s | 0.0139 s | 0.0123 s | 113000.0000 | 38000.0000 |  7000.0000 |    413 MB |
| ProcessIncrements |    15 |  True | 1.464 s | 0.0069 s | 0.0058 s | 178000.0000 | 48000.0000 | 15000.0000 |    611 MB |