@twop 2022-04-19
After the latest changes the "old" benchmarks for rapid updates has way better numbers:

### Before
|          Method | depth |       Mean |    Error |   StdDev |       Gen 0 |       Gen 1 |      Gen 2 | Allocated |
|---------------- |------ |-----------:|---------:|---------:|------------:|------------:|-----------:|----------:|
| ProcessMessages |    10 |   163.6 ms |  3.19 ms |  4.47 ms |  24000.0000 |   6333.3333 |  1000.0000 |     75 MB |
| ProcessMessages |    15 | 2,639.5 ms | 34.69 ms | 30.75 ms | 300000.0000 | 120000.0000 | 51000.0000 |    829 MB |

### After
|          Method | depth |       Mean |    Error |   StdDev |       Gen 0 |       Gen 1 |      Gen 2 | Allocated |
|---------------- |------ |-----------:|---------:|---------:|------------:|------------:|-----------:|----------:|
| ProcessMessages |    10 |   150.0 ms |  0.77 ms |  0.72 ms |  22750.0000 |   6000.0000 |  1000.0000 |     75 MB |
| ProcessMessages |    15 | 2,495.9 ms | 24.08 ms | 22.53 ms | 295000.0000 | 116000.0000 | 47000.0000 |    832 MB |


## What changed

Note that there is no difference In memory consumption but the access to `update` and `compare` functions and the functions themselves are more efficient.  In my guesstimate it went from somewhere 5-7 dynamic dispatches per attribute comparison to about 2-3