@twop 2022-04-19
Also I optimized diffing of attributes so much that it is almost indistinguishable from small scalars

|            Method | depth | boxed |    Mean |    Error |   StdDev |       Gen 0 |      Gen 1 |     Gen 2 | Allocated |
|------------------ |------ |------ |--------:|---------:|---------:|------------:|-----------:|----------:|----------:|
| ProcessIncrements |    15 | False | 1.202 s | 0.0102 s | 0.0090 s | 118000.0000 | 37000.0000 | 8000.0000 |    413 MB |
| ProcessIncrements |    15 |  True | 1.263 s | 0.0074 s | 0.0069 s | 119000.0000 | 42000.0000 | 8000.0000 |    479 MB |


BUT, it does consume significantly less memory which is probably going to be way more important for mobile devices