@twop 2022-05-04
I think I will have time this week.

SO we can do two things:
1. merge and then I convert XF attributes as a separate PR as a follow-up
2. I will take a stub at converting attributes as a part of this PR this week.

@TimLariviere what are your thoughts?